combine graph state and bell state modules into entangled state module. combine tests (which pass) add deprecation warning
